requirements.txt
python-telegram-bot==21.9
edge-tts==7.2.8
python-dotenv==1.0.1
